refactor: rename file to scraping_util
